move pre and during hooks t o ProxyLoggin
