+++ tests/mcp_tests/test_mcp_server.py
@@ -35,7 +35,7 @@ async def test_mcp_server_manager():
     print("TOOLS FROM MCP SERVER MANAGER== ", tools)
 
     result = await mcp_server_manager.call_tool(
-        name="gmail_send_email", arguments={"body": "Test"}
+        name="gmail_send_email", arguments={"body": "Test"}, proxy_logging_obj=None
     )
     print("RESULT FROM CALLING TOOL FROM MCP SERVER MANAGER== ", result)
 
@@ -105,6 +105,7 @@ async def test_mcp_server_manager_https_server():
                 "message": "Test",
                 "instructions": "Test",
             },
+            proxy_logging_obj=None,
         )
         print("RESULT FROM CALLING TOOL FROM MCP SERVER MANAGER== ", result)
         
@@ -248,7 +249,8 @@ async def test_mcp_http_transport_call_tool_mock():
                 "to": "test@example.com",
                 "subject": "Test Subject",
                 "body": "Test email body"
-            }
+            },
+            proxy_logging_obj=None,
         )
         
         # Assertions
@@ -308,7 +310,8 @@ async def test_mcp_http_transport_call_tool_error_mock():
         # Call the tool with invalid data
         result = await test_manager.call_tool(
             name="gmail_send_email",
-            arguments={"to": "invalid-email", "subject": "Test", "body": "Test"}
+            arguments={"to": "invalid-email", "subject": "Test", "body": "Test"},
+            proxy_logging_obj=None,
         )
         
         # Assertions for error case
@@ -343,7 +346,8 @@ async def test_mcp_http_transport_tool_not_found():
     with pytest.raises(ValueError, match="Tool nonexistent_tool not found"):
         await test_manager.call_tool(
             name="nonexistent_tool",
-            arguments={"param": "value"}
+            arguments={"param": "value"},
+            proxy_logging_obj=None,
         )
